# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Brainite API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Sources",
        "description": "List this workspace's source connections.",
        "operationId": "list_sources_api_v1_sources_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/sources/{provider}/authorize": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Authorize",
        "description": "Begin the OAuth flow: return the provider consent URL to redirect the user to.\n\nSubdomain-scoped providers (Zendesk) pass ``{\"subdomain\": \"acme\"}`` in the body.\n``returnTo`` (optional, allowlisted) picks the frontend path the callback redirects\nto \u2014 e.g. ``/onboarding`` \u2014 instead of the default sources page. The ``Origin``\nheader (allowlisted against ``FRONTEND_URLS``) picks which frontend origin it lands on.",
        "operationId": "authorize_api_v1_sources__provider__authorize_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/AuthorizeStartRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{provider}/callback": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Callback",
        "description": "OAuth redirect target: exchange the code, store the connection, bounce to the dashboard.\n\n``code`` is optional because a pure GitHub App install redirects here with an\n``installation_id`` and no ``code``. ``error`` is set when the user declines the\nconsent screen (e.g. ``error=access_denied``); the service redirects back cleanly\ninstead of trying to exchange a missing code.",
        "operationId": "callback_api_v1_sources__provider__callback_get",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "installation_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Installation Id"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}/report": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Read Report",
        "description": "What this source read, what became knowledge, and why the rest didn't.\n\nA successful import that yields no skills is indistinguishable from a broken one\nunless the discard reasons are visible \u2014 the pipeline records them per event, and\nthis is the surface that shows them.",
        "operationId": "read_report_api_v1_sources__source_id__report_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}/backfill": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Backfill",
        "description": "Import this source's history (a sweep scoped to one connection).\n\n``202`` for a newly started import, ``200`` when one covering this source was\nalready in flight \u2014 the same created/existing split as ``POST /sweeps``, and the\nsame ``SweepOut`` body, so a caller can poll ``GET /sweeps/{id}`` either way.",
        "operationId": "backfill_api_v1_sources__source_id__backfill_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}/disconnect": {
      "post": {
        "tags": [
          "sources"
        ],
        "summary": "Disconnect",
        "description": "Revoke (best-effort) and delete a source connection.",
        "operationId": "disconnect_api_v1_sources__source_id__disconnect_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/sources/{source_id}/channels": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Channels",
        "description": "Return the source's scope: discoverable channels + the persisted lookback window.",
        "operationId": "list_channels_api_v1_sources__source_id__channels_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "sources"
        ],
        "summary": "Select Channels",
        "description": "Persist the scope (channel selection + lookback) and echo back what was stored.",
        "operationId": "select_channels_api_v1_sources__source_id__channels_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChannelSelectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AuthorizeStartRequest": {
        "properties": {
          "subdomain": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subdomain"
          },
          "returnTo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returnto"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "AuthorizeStartRequest"
      },
...
      "ChannelSelectRequest": {
        "properties": {
          "channels": {
            "items": {
              "$ref": "#/components/schemas/ChannelSelection"
            },
            "type": "array",
            "title": "Channels"
          },
          "lookbackDays": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 730.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Lookbackdays"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "channels"
        ],
        "title": "ChannelSelectRequest"
      },
      "ChannelSelection": {
        "properties": {
          "externalId": {
            "type": "string",
            "title": "Externalid"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "selected": {
            "type": "boolean",
            "title": "Selected",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "externalId",
          "name"
        ],
        "title": "ChannelSelection"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}